Add the new implementation of the internal encoder of mrmime
This new encoder is not very different from the old one. Indeed, the biggest
update is about Level0/Enclosure which integrate an emitter to break out
continuation. Then, we are able to properly make a module like Format and a GADT
to encode any values.

In other side, Pretty (named Level1) is a much easier implementation of boxes
and pretty-printing which only fits under the purpose of mrmime - and emails in
general. With this new implementation, we are more able to produce well-formed
emails and tun behaviors of the prettier encoder more precisely.