Update all encoders to use the new interface.
It seems that a big change, but it's a minor one when it´s mostly a deletion of
identity function which is useless now and a rewrite of renamed boxes elements.

Mailbox, Content_type and MessageID are the only module where the diff is not
only a translation but where we put some [cut] elements in some places.