Fix a bug about the content-type parser which does not allow optional CFWS token between some characters.
According RFC 822, any spaces should be replaced by, at least, [FWS] token. This
is not the case before this commit in a Content-Type value. We fix the parser
where the new version of the encoder needs some [cut] token in some part to
split properly (and respect the 78 characters per line rule) the value.

The [cut] token is needed inside the Content-Type value encoder. Otherwise, it
generate:

Content-Type: text/plain; ... charset=
utf-8

And "utf-8" does not start with a proper whitespace.

So this commit adds some [C]?[FWS] token which surround '=' and ';' to be able
to parse what mrmime generate - this is NOT described by RFC 2045 explicitely.
This bug was reach by fuzzer.